@@ -12,28 +12,28 @@
   },
   "dependencies": {
     "classnames": "^2.2.4",
-    "react-input-autosize": "^0.6.13"
+    "react-input-autosize": "^1.1.0"
   },
   "devDependencies": {
-    "babel": "^5.8.23",
-    "babel-eslint": "^4.1.3",
+    "babel": "^6.5.2",
+    "babel-eslint": "^6.1.2",
     "chai": "^3.5.0",
     "coveralls": "^2.11.9",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.15.0",
+    "eslint": "^3.2.2",
+    "eslint-plugin-react": "^6.0.0",
     "gulp": "^3.9.1",
     "isomorphic-fetch": "^2.2.1",
     "istanbul": "^0.4.3",
-    "jsdom": "^8.4.1",
-    "mocha": "^2.4.5",
+    "jsdom": "^9.4.1",
+    "mocha": "^3.0.0",
     "react": "^15.0",
     "react-addons-shallow-compare": "^15.0.1",
     "react-addons-test-utils": "^15.0",
     "react-component-gulp-tasks": "^0.7.7",
     "react-dom": "^15.0",
     "react-gravatar": "^2.4.3",
-    "react-virtualized": "^6.3.0",
-    "react-virtualized-select": "^0.0.4",
+    "react-virtualized": "^7.17.1",
+    "react-virtualized-select": "^1.3.0",
     "sinon": "^1.17.3",
     "unexpected": "^10.13.2",
     "unexpected-dom": "^3.1.0",
