@@ -12,12 +12,12 @@ var MultiSelectValueList = _react2['default'].createClass({
 
   propTypes: {
     children: _react2['default'].PropTypes.node,
+    className: _react2['default'].PropTypes.string,
     disabled: _react2['default'].PropTypes.bool, // disabled prop passed to ReactSelect
     id: _react2['default'].PropTypes.string, // Unique id for the value - used for aria
     onClick: _react2['default'].PropTypes.func, // method to handle click on value label
     onRemove: _react2['default'].PropTypes.func, // method to handle removal of the value
-    value: _react2['default'].PropTypes.object.isRequired, // the option object for this value
-    className: _react2['default'].PropTypes.string
+    value: _react2['default'].PropTypes.object.isRequired // the option object for this value
   },
 
   handleMouseDown: function handleMouseDown(event) {
@@ -66,7 +66,7 @@ var MultiSelectValueList = _react2['default'].createClass({
       return;
     }
     return _react2['default'].createElement('span', {
-      style: { float: "right" },
+      style: { float: 'right' },
       className: 'icon-cross3',
       'aria-hidden': 'true',
       onMouseDown: this.onRemove,
@@ -78,7 +78,7 @@ var MultiSelectValueList = _react2['default'].createClass({
   renderLabel: function renderLabel() {
     return _react2['default'].createElement(
       'span',
-      { style: { width: "95%" } },
+      { style: { width: '95%' } },
       this.props.children
     );
   },
@@ -87,7 +87,7 @@ var MultiSelectValueList = _react2['default'].createClass({
     return _react2['default'].createElement(
       'div',
       {
-        className: "Select--multiItem " + this.props.className,
+        className: 'Select--multiItem ' + this.props.className,
         title: this.props.value.title },
       this.renderLabel(),
       this.renderRemoveIcon()
